Use custom hook for pointer tracker
